Abstract class, and additional functions and data
In order to use sort, algorithm library was included. Added comments to better clarify...

Building class serves as a blueprint for airports and other buildings.

Added a data member in the airPort class (airName) for the name of airport ("Mineta" airport in San Jose), along with associated functions. 

Created a display function for mapQueue object along with sorting functions (by name, city and air code).