Update World_Planes.cpp
Defined a function for Building class, changed constructors to reflect data member inclusion (airName)

Added noexcept to the move constructor (this might be unnecessary or causing a bug)

Renamed few functions to remove confusion between cityName and airName. 

freeAirSpace function was created to point a pointer to null to prevent memory leak. (STILL NOT WORKING, when copied/moved airPort pointers are involved, issues arise)

sort functions are defined

Note: there are commented out code that might or might not be used later