Aggregate sort and macros defined
In the source.cpp,  sort and display functions  are aggregated into one.

In the header file, macros are defined for sorting purposes, and a prototype was created for aggregate sort function.

In the cpp file, the aggregate sort function is defined

CAUTION: Move constructor and its operator are defined, but not sure if it actually works....